fix: Navbar 스크롤시 투명색으로 변하게 수정 및 <Navbar/> 호출 위치 변경
